fixed auth0 and made nav bar conditional
