+++ src/Nav/nav.js
@@ -1,8 +1,9 @@
-import {
-    Link
-  } from "react-router-dom";
+import { useAuth0 } from "@auth0/auth0-react";
+import LoggedOut from "../LoggedOut/loggedout";
+import LoggedIn from "../LoggedIn/loggedin";
   
   function Nav() {
+    const { isAuthenticated } = useAuth0();
     return (
       <nav className="flex nav align-top container mx-auto bg-dark h-16">
         <div className="flex flex-1 items-center justify-center sm:items-stretch sm:justify-start">
@@ -11,14 +12,7 @@ import {
           </div>
           <div className="hidden sm:ml-6 sm:block">
             <div className="flex space-x-4">
-              <Link to="/landing" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Landing</Link>
-              <Link to="/" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Dashboard</Link>
-              <Link to="/organization" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Organization</Link>
-              <Link to="/profile" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Profile</Link>
-              <Link to="/recent" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">Recent</Link>
-              <Link to="/login" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">LogIn</Link>
-              <Link to="/logout" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">LogOut</Link>
-              <Link to="/signup" className="text-slate-300 hover:bg-slate-700 hover:text-white rounded-md px-3 py-2 text-sm font-medium">SignUp</Link>
+              {isAuthenticated ? <LoggedIn/>: <LoggedOut/>}
             </div>
           </div>
         </div>
